[Caleb, Rob] 3/5/2023 changes
Made ArmToPosition work. made ArmElbowToPosition use an on target count. Added a ArmMoveShoulderElbow which is a command that moves the shoulder of the arm. Made are move to position depricated. removed ArmResetElbowToZero. Moved ArmResetShoulderToZero. Made an end condition for ArmShoulderToPosition.  Added a ArmPositionSupplier.  Added constants to constants. Changed key bindings in robot container. Fixed pid on the arm. Made goal locations store ArmPositions insted of the height of the poles. Started working on rumble for intake.